feat(api): Add data model for recent searches, and endpoint to list them (SEN-350, SEN-349)
This adds a data model to capture a user's recent searches in an organization. We support a `type`
column so that this can be used for issue/event/other searches as needed.

Also includes the api endpoint to list the searches. Endpoint for creating them will be in a follow
up diff.